refactor into traces_core
