+++ magpylib/_src/obj_classes/class_current_Line.py
@@ -1,5 +1,5 @@
 """Line current class code"""
-from magpylib._src.display.traces_generic import make_Line
+from magpylib._src.display.traces_core import make_Line
 from magpylib._src.fields.field_BH_line import current_vertices_field
 from magpylib._src.input_checks import check_format_input_vertices
 from magpylib._src.obj_classes.class_BaseExcitations import BaseCurrent
