+++ magpylib/_src/obj_classes/class_magnet_Cuboid.py
@@ -1,5 +1,5 @@
 """Magnet Cuboid class code"""
-from magpylib._src.display.traces_generic import make_Cuboid
+from magpylib._src.display.traces_core import make_Cuboid
 from magpylib._src.fields.field_BH_cuboid import magnet_cuboid_field
 from magpylib._src.input_checks import check_format_input_vector
 from magpylib._src.obj_classes.class_BaseExcitations import BaseMagnet
