+++ magpylib/_src/obj_classes/class_magnet_Cylinder.py
@@ -1,5 +1,5 @@
 """Magnet Cylinder class code"""
-from magpylib._src.display.traces_generic import make_Cylinder
+from magpylib._src.display.traces_core import make_Cylinder
 from magpylib._src.fields.field_BH_cylinder_segment import magnet_cylinder_field
 from magpylib._src.input_checks import check_format_input_vector
 from magpylib._src.obj_classes.class_BaseExcitations import BaseMagnet
